uos_api/requirements.txt
psycopg[binary]
python-dotenv
Flask

